first draft: eslint-module-utils as a separate package
