+++ .travis.yml
@@ -1,7 +1,7 @@
 language: node_js
 node_js:
   # - 0.10
-  - 0.12
+  # - 0.12
   - 4
   - stable
