+++ src/rules/no-deprecated.js
@@ -1,5 +1,5 @@
-import Exports from '../core/getExports'
-import declaredScope from '../core/declaredScope'
+import Exports from 'eslint-module-utils/ExportMap'
+import declaredScope from 'eslint-module-utils/declaredScope'
 
 module.exports = function (context) {
   const deprecated = new Map()
