+++ src/rules/no-unresolved.js
@@ -3,7 +3,7 @@
  * @author Ben Mosher
  */
 
-import resolve from '../core/resolve'
+import resolve from 'eslint-module-utils/resolve'
 
 module.exports = function (context) {
